[libc++][lit] Atomically update the persistent cache (#66538)
When running multiple shards in parallel, one shard might write to the
cache while another one is reading this cache. Instead of updating the
file in place, write to a temporary file and swap the cache file using
os.replace(). This is an atomic operation and means shards will either
see the old state or the new one.

NOKEYCHECK=True
GitOrigin-RevId: 14882d6b7440e797d8d60cb1f8207fe332cae033