[libc++] Use correct size for deallocation of arrays in shared_ptr (#68233)
Fixes #68051.

Current implementation passes the number of `_AlignedStorage` objects
when it calls to `allocate` and the number of **bytes** on `deallocate`.
This only applies to allocations that allocate control block and the
storage together, i.e. `make_shared` and `allocate_shared`.

Found by ASan under Clang combined with `-fsized-deallocation`.

NOKEYCHECK=True
GitOrigin-RevId: f722db02d359e29ca001b78197ee1a275f8c3d7c