[libc++][test] disable all atomic<long double> tests (#74201)
Clang's support for atomic operations on long doubles is currently broken,
so these tests don't work everywhere. This is a long standing condition and
the goal of this patch is to get the CI green again on all platforms.

The actual Clang fixes will be pursued separately.

Fixes #73791

NOKEYCHECK=True
GitOrigin-RevId: 66be0571932c016d5904fa9e14423b4e0a3fd586