[libc++] Fix incomplete user-defined ctype specialization in test (#74630)
The specialization was non-conforming because it was missing a bunch of
member functions. Those were missing probably just as an oversight
coupled with a bit of laziness -- the rule that user-defined
specializations need to match the base template is usually OK to take
with a grain of salt, but not when the code is supposed to be portable,
which our test suite aims to be.

Fixes #74214

NOKEYCHECK=True
GitOrigin-RevId: 785e0945ce49cf5be0d8fcc0874ad6a5c94fc920