Reapply "[Sema] Fix crash on invalid code with parenthesized aggregate initialization" (#76272)
With updates the libc++ tests.

This reverts commit 2205d2334f3c859ad9f6c65ed950bfb3bb6f7cbe and relands
86dc6e15f22610bbb53eb4efda0a178ecefc933a and
7ab16fb5207fe187ab999f882069bd632d2e68e5.

Original commit was reverted because of failing libc++ tests, see #76232 for
the discussion.

The errors in the tests are spurious in the first place (coming from initialization
of invalid classes), so update the tests to match new behavior that does
not show those errors.

NOKEYCHECK=True
GitOrigin-RevId: 02347fc7191ff4d073f439dde6523add3f5496de