[libc++][docs] Document the libc++ Lit testing format naming scheme (#73136)
As a new contributor, I found it hard to find the documentation for the
meaning of the names of different tests and how those names translate to
Lit. This patch moves the documentation to the RST documentation we
publish on the website instead of leaving it in the source code only.

NOKEYCHECK=True
GitOrigin-RevId: d301539b777c5047d1420003b4ab1e05a1f87166