[libc++] Fix Coverity warning about use-after-move (#78780)
While the code is technically correct because the index is never
actually moved from (and anyway that wouldn't matter since it's an
integer), it's still better style not to access an object after it has
been moved-from. Since this is so easy to do, just save the index in a
temporary variable.

rdar://120501577

NOKEYCHECK=True
GitOrigin-RevId: 86b6dfc619695b5bfb1880b2ed9abb4a6805fbe6